refactor: rename login to auth
