+++ androidApp/src/main/java/com/example/tasky/android/MainNavHost.kt
@@ -6,8 +6,8 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import com.example.tasky.android.agenda.AgendaGraph
 import com.example.tasky.android.agenda.agendaGraph
-import com.example.tasky.android.login.LoginGraph
-import com.example.tasky.android.login.loginGraph
+import com.example.tasky.android.auth.LoginGraph
+import com.example.tasky.android.auth.loginGraph
 
 @Composable
 fun MainNavHost(
