+++ androidApp/src/main/java/com/example/tasky/android/auth/viewmodel/RegisterViewModel.kt
@@ -1,9 +1,9 @@
-package com.example.tasky.android.login.viewmodel
+package com.example.tasky.android.auth.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.tasky.android.login.screen.RegisterScreenEvent
-import com.example.tasky.android.login.screen.RegisterScreenState
+import com.example.tasky.android.auth.screen.RegisterScreenEvent
+import com.example.tasky.android.auth.screen.RegisterScreenState
 import com.example.tasky.auth.domain.IAuthRepository
 import com.example.tasky.auth.domain.util.Validator
 import com.example.tasky.common.domain.model.onError
